modified:   apps/build_pcd_file_settings.xml 	modified:   apps/build_pcd_files.py 	modified:   apps/component_test/main.cpp 	modified:   apps/test_python.py 	modified:   common/KinectInterface.cpp 	modified:   common/KinectInterface.h 	modified:   common/MappingInterface.cpp 	modified:   common/SLAMInterface.cpp 	modified:   common/SLAMInterface.h 	modified:   common/VisualOdometryInterface.cpp 	modified:   common/VisualOdometryInterface.h 	modified:   gui/LogicBridge.cpp 	modified:   gui/PointCloudViewer.cpp 	modified:   gui/RobXControl.cpp 	modified:   interfaces/FovisInterface.cpp 	modified:   interfaces/ICPInterface.cpp 	modified:   interfaces/KinectDatasetReader.cpp 	modified:   interfaces/KinectDatasetReader.h 	modified:   interfaces/KinectDeviceReader.cpp 	modified:   interfaces/KinectDeviceReader.h 	modified:   interfaces/iSAM2Interface.cpp 	modified:   interfaces/iSAM2Interface.h 	modified:   libksrobot/libksrobot_common.cpp 	modified:   libksrobot/libksrobot_eigen.cpp

+++ common/KinectInterface.h
@@ -18,6 +18,12 @@ namespace common
 
 typedef pcl::PointCloud<pcl::PointXYZRGBA>                                  KinectPointCloud;
 
+// #ifdef KSROBOT_USE_DOUBLE_PRECISION_MATH
+//     typedef Eigen::Isometry3d       IsometryTransform
+// #else
+//     typedef Eigen::Isometry3f       IsometryTransform
+// #endif //KSROBOT_USE_DOUBLE_PRECISION_MATH
+
 class KinectInterface : public Interface
 {
 public:
@@ -58,7 +64,7 @@ public:
     inline float                                UnitsPerMeter() const { return 0; }
     
     virtual bool                                ProvidesGroundTruth() = 0;
-    virtual Eigen::Isometry3d                   GetCurrentGroundTruth() = 0;
+    virtual Eigen::Isometry3f                   GetCurrentGroundTruth() = 0;
     
     inline void                                 EnablePointCloudGeneration(bool enable);
     inline void                                 EnableFloatDepthGeneration(bool enable);
