+++ interfaces/ICPInterface.cpp
@@ -69,8 +69,8 @@ bool ICPInterface::RunSingleCycle()
     
     mOdomTimer->Start();
         mImpl->ICP.align(mAlignedCloud);
-        //mMotionEstimate = mImpl->ICP.getFinalTransformation().cast<double>();
-        mMotion->MotionEstimate = mImpl->ICP.getFinalTransformation().cast<double>();
+        //mMotionEstimate = mImpl->ICP.getFinalTransformation();
+        mMotion->MotionEstimate = mImpl->ICP.getFinalTransformation();
     mOdomTimer->Stop();
     
     FinishCycle();
